rectangles:  Make skia and cairo test cases consistent
Both tests now produce identical output, so this is now providing a
solid apples-to-apples comparison.  This shows cairo trumping skia's
performance by about 300% on the image backend.

Note that I disabled antialiasing in the cairo case in order to make the
rectangle lines draw as sharply as they do in skia.  It looks like skia
may have antialiasing off by default; to be certain I've forced it off.