cmake: Make GLX/EGL compilation optional (with EGL as default)
GLX and GLES2/3 are incompatible as apparently the have overlapping
API's so when both are linked in an executable one lib's functions are
going to subvert the others.

This patch provides options to configure GLX and EGL on or off.

Longer term, this will probably need to be revisited.  Perhaps there
should be a GL_FLAVOR, that can be set to "glx", "egl", etc. so only one
is defined at a time.  Or, perhaps we generate multiple binaries, one
for each, ala caskbench_egl, caskbench_glx, etc.