Rename Shapes enum to shape_type_t and fix enum names
We already have a 'shapes' data structure, so to avoid confusion don't
name our shape type enums 'Shapes'.  Instead, name it shape_type_t,
which is more consistent with the code style of other enums in the
codebase.

Also, rename Circle / Triangle / etc. to CB_SHAPE_CIRCLE /
CB_SHAPE_TRIANGLE / etc.  The former could be confused as class names;
the latter are less likely to get confused this way, and follow the
style of other enums in the codebase.