utils: Add sample image mandel.png
This is a copy of the test image used in flashcanvas.  I downloaded it
from:

  http://flashcanvas.net/examples/dl.dropbox.com/u/1865210/mindcat/images/mandel.png

Near as I can tell, this is a cropped and scaled down version of
one of Wikipedia's mandelbrot set images, specifically:

  http://commons.wikimedia.org/wiki/File:Mandel_zoom_07_to_08.png

Licensing of this image is covered by Creative Commons Attribution-Share
Alike 3.0.  If this assumption is in error, please let us know, so we
can  replace this copy with one taken directly from wikipedia.